Stop appending final newlines
The behavior regarding final newlines has changed between DjHTML
v1.4.14 and v1.5.0. It used to always append the final newline,
but now this will only happen when the source file already
contains a final newline.

See https://github.com/rtts/djhtml/issues/56 for the discussion
that led to this change.

If you still need the old behavior, the best option is to use
https://github.com/pre-commit/pre-commit-hooks#end-of-file-fixer
by adding the following to your `.pre-commit-config.yaml`:

    - repo: https://github.com/pre-commit/pre-commit-hooks
      rev: main  # replace with the latest tag on GitHub
      hooks:
        - id: end-of-file-fixer

Closes #56