Support passing directories as command-line arguments
This makes it easier to run djhtml over a bunch of files in a
directory or directories.

The default suffixes are used as a filter with no configuration
allowed (as black seems to work).

The existing functionality whereby file names with any prefix can be
specified will continue to work.