JavaScript regular expression literals
There is still one broken case that is _very_ hard to fix:

    var re = x / y // comment

Here the string "/ y /" is recognized as a literal while it is not. This
means the comment is not tokenized as a comment, and if it contains any
opening tokens the next line will be incorrectly indented. A working example
is provided in the unittests.