Compare actual with expected tokenization in unittests
This guarantees that DjHTML is exactly behaving as expected, even if a
tokenization mishap doesn't result in changed indentation. An added bonus is
that by tracking these files in Git, we will notice changes in tokenization
immediately.

The initial token files were generated by the supplied `generate_tokens.py`
script.