Relative multi-line HTML elements
Reintroduces "relative multi-line" indentation to DjHTML:

    <long-html-tag
        attribute1="value"
        attribute2="value"
        attribute3="value"/>

However, this only happens when the tag name is followed by a newline.
Otherwise, "absolute multi-line" indentation will be used:

    <long-html-tag attribute1="value"
                   attribute2="value"
                   attribute3="value"/>

This is a good rule, but also the last rule I'm willing to add regarding
multi-line indentation. Our users deserve a stable set of indentation
principles, and this is it. No more bikeshedding.