Compromise on JS regular expression literals
The compromise is that they cannot contain spaces. If they can, it
becomes impossible to distinguish between

    re = / foo( // comment

and

    not_re = x / foo( // comment

In other words, it's impossible to distinguish between the division
sign and the beginning of a regex literal. It still is, but now it
will be less of problem for literals without spaces.