Support passing non-standard extra blocks.
This enables support for handling `{% nested_widget %}...{% endnested %}` blocks used by sniplates.

This creates an extra command-line argument (-b --extra-block), that can be supplied multiple times. Values like "nested_widget,endnested" are split into a tuple, and then passed as a list of tuples into the parser, which then prefers to use this for detecting an end tag, but falls back to the previous mechanism if the current tag does not have an explicit end-tag value.